test: split unit and integration tests
