+++ pkg/detectors/maxmindlicense/v1/maxmindlicense_integration_test.go
